the testform and testqml cases were basically the same now, so lets just merge em
