+++ QMLComponents/components/JASP/Controls/AssignedVariablesList.qml
@@ -15,6 +15,7 @@
 // License along with this program.  If not, see
 // <http://www.gnu.org/licenses/>.
 //
+import JASP
 
 VariablesList
 {
