+++ Tests/testqml.cpp
@@ -1,7 +1,11 @@
-#include "testqml.h"
+#include <iostream>
 #include <QQmlEngine>
 #include <QQmlContext>
+#include "tempfiles.h"
 #include "jasptheme.h"
+#include "processinfo.h"
+#include "testqml.h"
+#include "datasetprovider.h"
 #include "preferencesmodelbase.h"
 #include "utilities/qmlutils.h"
 
@@ -18,6 +22,11 @@ void TestQml::applicationAvailable()
 
 void TestQml::qmlEngineAvailable(QQmlEngine *engine)
 {
+
+	TempFiles::init(ProcessInfo::currentPID());
+
+	TempFiles::clearSessionDir();
+
 	// Initialization requiring the QQmlEngine to be constructed
 	engine->rootContext()->setContextProperty("myContextProperty", QVariant(true));
 
@@ -38,12 +47,20 @@ void TestQml::qmlEngineAvailable(QQmlEngine *engine)
 	engine->rootContext()->setContextProperty("jaspTheme",			_theme);
 	engine->rootContext()->setContextProperty("preferencesModel",	_prefs);
 
+	_prov = DataSetProvider::getProvider(false, false);
+
+	_prov->absorbInfo(VariableInfo::DataSetValues, "TestColumn", 0, QVariantList({1,2,3,4,5}));
 }
 
 void TestQml::cleanupTestCase()
 {
-}
+	delete _theme;
+	delete _prov;
 
+	_theme = nullptr;
+	_prefs = nullptr;
+	_prov  = nullptr;
+}
 
 QUICK_TEST_MAIN_WITH_SETUP(qmltest, TestQml);
