+++ Tests/testqml.h
@@ -4,6 +4,7 @@
 #include <QtQuickTest>
 
 class PreferencesModelBase;
+class DataSetProvider;
 class QQmlEngine;
 class JaspTheme;
 
@@ -23,6 +24,7 @@ public slots:
 private:
 	JaspTheme				* _theme = nullptr;
 	PreferencesModelBase	* _prefs = nullptr;
+	DataSetProvider			* _prov  = nullptr;
 };
 
 #endif // TESTQML_H
